refactor(core): move the velesql/, filter/, wire/ and api_types/ inline test modules to sibling files (#1918)
Third of four lots for velesdb-core. Eleven files, all mono-block and at
end of file; the only judgment calls are naming.

- cache.rs and error.rs already have a <stem>_tests.rs sibling, so their
  inline unit tests land in cache_unit_tests.rs and error_unit_tests.rs.
- filter/mod.rs names its module condition_type_names_tests; the name is
  kept and the sibling named after it.
- velesql/ast/mod.rs takes ast_tests.rs, named for its directory rather
  than the generic mod_tests.rs.

The sorted test-name list is byte-identical before and after, re-checked
after rustfmt: 5357 tests, 5341 passing and 16 pre-existing ignored.
Baseline down to 42 files.
